refactor(evm-explorer): split transfer translation modules
Why:
- keep the EVM explorer adapter shape aligned with repo adapter guidance
- separate parsing and mapping seams so translation.py stays a dispatcher

What:
- move translation context into models.py and parsing helpers into parsing.py
- split native and token transfer rules into bounded modules
- keep issue construction and dispatcher behavior local to the adapter package

Checks:
- UV_PROJECT_ENVIRONMENT="$HOME/.venvs/tallylot-py312" uv run pytest --no-cov src/tallylot/adapters/sources/explorers/evm_explorer/tests/test_adapter.py tests/contract/test_pylint_contracts.py -q
- UV_PROJECT_ENVIRONMENT="$HOME/.venvs/tallylot-py312" uv run python -m tools.run_quality_gates --full-tests
